requirements.txt
waitress
flask
Wikidata
Google-Search-API